requirements.txt
flask
sqlalchemy>=1.4
pymysql
azure-identity==1.23.0
azure-keyvault-secrets==4.9.0

# delete in production
# python-dotenv